Fix (MSVC) compiler warnings reg. `size_t` vs. `int`
as per code review, https://github.com/musescore/MuseScore/pull/11197#pullrequestreview-943197231

+++ src/engraving/libmscore/barline.h
@@ -98,7 +98,7 @@ public:
     // Score Tree functions
     EngravingObject* scanParent() const override;
     EngravingObject* scanChild(int idx) const override;
-    int scanChildCount() const override;
+    size_t scanChildCount() const override;
 
     BarLine* clone() const override { return new BarLine(*this); }
     Fraction playTick() const override;
